refactor: Move Artist from Containers to Scenes
(cherry picked from commit f699e43c7b904f066ed68cd6b5ef5d05fd49440d)
